pyproject.toml
[tool.poetry]
name = "djangopoetrysample"
version = "0.1.0"
description = "Sample project for django with poetry"
authors = ["Sachintha Dhananjana"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
flake8 = "^7.0.0"
django-debug-toolbar = "^4.4.2"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
